feat: refine knowledge base list document api
