+++ backend/app/rag/chat_config.py
@@ -1,5 +1,4 @@
 import os
-import json
 import logging
 from typing import Optional
